requirements.txt
selenium==4.28.0
pandas==2.2.3
numpy==2.2.2
openpyxl==3.1.5
requests==2.32.3
Pillow==11.1.0
pdfplumber==0.11.5
SQLAlchemy==2.0.37
cairosvg==2.7.1
beautifulsoup4==4.12.3
psycopg2-binary==2.9.10
lxml==5.3.0
google-auth==2.37.0
google-api-python-client==2.159.0
chardet==5.2.0
charset-normalizer==3.4.1
python-dotenv==1.0.1
pyvirtualdisplay==3.0
